+++ src/test/java/com/dreamypatisiel/devdevdev/test/MySQLTestContainer.java
@@ -1,0 +1,39 @@
+package com.dreamypatisiel.devdevdev.test;
+
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+/**
+ * MySQL 테스트컨테이너를 제공하는 공통 클래스
+ * 1. 테스트 클래스에서 이 클래스를 상속받거나
+ * 2. @ExtendWith(MySQLTestContainer.class) 어노테이션을 사용
+ */
+@Testcontainers
+public abstract class MySQLTestContainer {
+
+    @Container
+    @ServiceConnection
+    protected static MySQLContainer<?> mysql = new MySQLContainer<>("mysql:8.0")
+            .withDatabaseName("devdevdev_test")
+            .withUsername("test")
+            .withPassword("test")
+            .withCommand(
+                "--character-set-server=utf8mb4", 
+                "--collation-server=utf8mb4_unicode_ci", 
+                "--ngram_token_size=1"
+            );
+
+    @DynamicPropertySource
+    static void configureProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", mysql::getJdbcUrl);
+        registry.add("spring.datasource.username", mysql::getUsername);
+        registry.add("spring.datasource.password", mysql::getPassword);
+        registry.add("spring.jpa.database-platform", () -> "org.hibernate.dialect.MySQLDialect");
+        registry.add("spring.jpa.hibernate.ddl-auto", () -> "create-drop");
+        registry.add("spring.jpa.show-sql", () -> "true");
+    }
+}
